[SSN-101] feat: 루트 접속시 홈으로 리다이렉트 수정 및 헤더 수정
