+++ src/app/home/components/Banner.tsx
