+++ src/app/home/components/TrendAnalysisSection.tsx
