+++ src/app/home/components/hero/HeroCarousel.tsx
