+++ src/app/home/components/hero/HeroSection.tsx
