+++ src/app/home/components/hero/slides/DaehongSlide.tsx
