+++ src/app/home/components/hero/slides/HeroMainSlide.tsx
