+++ src/app/home/components/hero/slides/HsadSlide.tsx
